requirements.txt
pyyaml>=6.0
tqdm>=4.60.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
accelerate>=0.20.0